+++ src/prefect/runner/server.py
@@ -8,6 +8,7 @@ from prefect._vendor.fastapi.responses import JSONResponse
 from typing_extensions import Literal
 
 from prefect._internal.pydantic import HAS_PYDANTIC_V2
+from prefect._internal.schemas.validators import validate_values_conform_to_schema
 from prefect.client.orchestration import get_client
 from prefect.exceptions import MissingFlowError, ScriptError
 from prefect.flows import Flow, load_flow_from_entrypoint, load_flows_from_script
@@ -24,7 +25,6 @@ from prefect.settings import (
     PREFECT_RUNNER_SERVER_PORT,
 )
 from prefect.utilities.asyncutils import sync_compatible
-from prefect.utilities.validation import validate_values_conform_to_schema
 
 if TYPE_CHECKING:
     from prefect.client.schemas.responses import DeploymentResponse
