+++ src/prefect/server/api/validation.py
@@ -3,7 +3,7 @@ from typing import Union
 from prefect._vendor.fastapi import HTTPException, status
 
 import prefect.server.schemas as schemas
-from prefect.utilities.validation import validate_values_conform_to_schema
+from prefect._internal.schemas.validators import validate_values_conform_to_schema
 
 
 def _get_base_config_defaults(base_config: dict):
