+++ src/prefect/server/schemas/actions.py
@@ -19,6 +19,12 @@ else:
 
 import prefect.server.schemas as schemas
 from prefect._internal.compatibility.experimental import experimental_field
+from prefect._internal.schemas.validators import (
+    raise_on_name_alphanumeric_dashes_only,
+    raise_on_name_alphanumeric_underscores_only,
+    validate_schema,
+    validate_values_conform_to_schema,
+)
 from prefect.server.utilities.schemas import get_class_fields_only
 from prefect.server.utilities.schemas.bases import PrefectBaseModel
 from prefect.server.utilities.schemas.fields import DateTimeTZ
@@ -27,15 +33,7 @@ from prefect.server.utilities.schemas.transformations import (
     FieldFrom,
     copy_model_fields,
 )
-from prefect.server.utilities.schemas.validators import (
-    raise_on_name_alphanumeric_dashes_only,
-    raise_on_name_alphanumeric_underscores_only,
-)
 from prefect.utilities.templating import find_placeholders
-from prefect.utilities.validation import (
-    validate_schema,
-    validate_values_conform_to_schema,
-)
 
 
 def validate_block_type_slug(value):
